fix(webui): render context messages, hub/tenant lists and user echo in shared renderer
- Extract convertMossMessagesToTMessages from desktop RemoteConversationProvider
  into @sudowork/common/chatLib; webui server getContext now converts moss
  messages to TMessage so user/AI bubbles render
- Add assistant-hub.fetch-assistants and skill-hub.fetch-skills handlers in
  mossAdapter, routing tenant-exclusive lists via /tenant and hub lists with
  cursor/category/search params
- Drop isElectronDesktop branches in agents/skills pages and remove
  fetchSkillsHttp; both ends now go through the bridge invoke
- Pass organization.id as enterprise_code in the web session so the
  exclusive tab passes the enterprise gate
- Synthesize a user_content frame after send so the user's own bubble shows
  (no optimistic insert in the shared renderer)
- Show the back-to-mode-select link only on desktop
- Remove the outdated ui-reuse-map doc and stale console wording
