+++ apps/desktop/src/process/providers/RemoteConversationProvider.ts
@@ -7,6 +7,7 @@
 import type { ICreateConversationParams, IBridgeResponse, ISendMessageParams } from '@sudowork/host-bridge/ipcBridge';
 import type { TChatConversation } from '@sudowork/common/storage';
 import type { TMessage } from '@sudowork/common/chatLib';
+import { convertMossMessagesToTMessages } from '@sudowork/common/chatLib';
 import { isRemoteContainerPath } from '@sudowork/common/utils/workspaceSkillSync';
 import { getDatabase } from '@process/database';
 import WorkerManage from '@process/WorkerManage';
@@ -459,309 +460,6 @@ export class RemoteConversationProvider implements IConversationProvider {
   // ========== Messages / 消息 ==========
 
   /**
-   * Convert Moss Server messages to TMessage format (reusable)
-   * 将 Moss Server 消息转换为 TMessage 格式（可复用）
-   */
-  private convertMossMessagesToTMessages(allMessages: any[], conversationId: string, mossSessionId: string): { messages: TMessage[]; foundModel: string } {
-    const messages: TMessage[] = [];
-    const finishedMessageStatus = 'finish';
-    let messageIndex = 0;
-    let foundModel = '';
-
-    for (const msg of allMessages) {
-      const msgType = msg.type;
-      const innerRole = msg.message?.role;
-      const msgModel = msg.message?.model;
-      const timestamp = new Date(msg.timestamp || Date.now()).getTime();
-
-      if (msgType === 'assistant' && msgModel) {
-        foundModel = msgModel;
-      }
-
-      if (msgType === 'tool_use') {
-        const toolCallId = msg.tool_use_id || msg.id || msg.uuid || `${conversationId}-${messageIndex}`;
-        const responseToolCallId = msg.uuid || toolCallId;
-        const toolName = msg.name || 'Tool';
-        const rawInput = this.parseMossToolInput(msg.input);
-
-        if (toolName === 'AskUserQuestion') {
-          const question = typeof rawInput.question === 'string' ? rawInput.question : '';
-          const description = typeof rawInput.description === 'string' ? rawInput.description : undefined;
-          const options = Array.isArray(rawInput.options) ? rawInput.options.filter((option): option is string => typeof option === 'string') : [];
-
-          messages.push({
-            id: `${conversationId}-${messageIndex++}`,
-            msg_id: toolCallId,
-            conversation_id: conversationId,
-            type: 'acp_question',
-            position: 'left',
-            content: {
-              question,
-              intro: description,
-              options,
-              conversationId,
-              toolCallId,
-              responseToolCallId,
-            },
-            create_time: timestamp,
-            status: finishedMessageStatus,
-          } as unknown as TMessage);
-        } else {
-          messages.push({
-            id: `${conversationId}-${messageIndex++}`,
-            msg_id: toolCallId,
-            conversation_id: conversationId,
-            type: 'acp_tool_call',
-            position: 'left',
-            content: {
-              sessionId: mossSessionId,
-              update: {
-                sessionUpdate: 'tool_call',
-                toolCallId,
-                status: 'completed',
-                title: toolName,
-                kind: this.getMossToolKind(toolName),
-                rawInput,
-                content: [],
-              },
-            },
-            create_time: timestamp,
-            status: finishedMessageStatus,
-          } as unknown as TMessage);
-        }
-        continue;
-      }
-
-      if (msgType !== 'user' && msgType !== 'assistant' && innerRole !== 'user' && innerRole !== 'assistant') {
-        continue;
-      }
-
-      const contentArray = msg.message?.content || msg.content || [];
-      const isError = msg.error || msg.isApiErrorMessage;
-      const msgRole = msgType || innerRole || 'unknown';
-
-      if (isError && msgRole === 'assistant') {
-        const errorText = Array.isArray(contentArray)
-          ? contentArray
-              .filter((c: any) => c?.type === 'text')
-              .map((c: any) => c.text || '')
-              .join('\n')
-          : typeof contentArray === 'string'
-            ? contentArray
-            : '';
-        messages.push({
-          id: `${conversationId}-${messageIndex++}`,
-          conversation_id: conversationId,
-          type: 'tips',
-          position: 'left',
-          content: { content: errorText || msg.error || 'Unknown error', type: 'error' },
-          create_time: timestamp,
-          status: finishedMessageStatus,
-        } as unknown as TMessage);
-        continue;
-      }
-
-      if (msgRole === 'user' && Array.isArray(contentArray)) {
-        for (const block of contentArray) {
-          if (block?.type === 'text') {
-            const textContent = this.extractDisplayUserContent(block.text || '');
-            if (textContent && textContent.trim()) {
-              messages.push({
-                id: `${conversationId}-${messageIndex++}`,
-                conversation_id: conversationId,
-                type: 'text',
-                role: 'user',
-                position: 'right',
-                content: { content: textContent },
-                create_time: timestamp,
-                status: finishedMessageStatus,
-              } as unknown as TMessage);
-            }
-          } else if (block?.type === 'tool_result') {
-            const toolUseId = block.tool_use_id || block.toolCallId;
-            const isError = block.is_error;
-            const resultContent = block.content || '';
-
-            if (isError) {
-              messages.push({
-                id: `${conversationId}-${messageIndex++}`,
-                msg_id: toolUseId,
-                conversation_id: conversationId,
-                type: 'tips',
-                position: 'left',
-                content: { content: resultContent || 'Tool execution failed', type: 'error' },
-                create_time: timestamp,
-                status: finishedMessageStatus,
-              } as unknown as TMessage);
-            } else {
-              messages.push({
-                id: `${conversationId}-${messageIndex++}`,
-                msg_id: toolUseId,
-                conversation_id: conversationId,
-                type: 'acp_tool_call',
-                position: 'left',
-                content: {
-                  sessionId: mossSessionId,
-                  update: {
-                    sessionUpdate: 'tool_call_update',
-                    toolCallId: toolUseId,
-                    status: 'completed',
-                    content: [{ type: 'content', content: { type: 'text', text: resultContent } }],
-                  },
-                },
-                create_time: timestamp,
-                status: finishedMessageStatus,
-              } as unknown as TMessage);
-            }
-          }
-        }
-        continue;
-      }
-
-      if (msgRole === 'user') {
-        const textContent = Array.isArray(contentArray)
-          ? contentArray
-              .filter((c: any) => c?.type === 'text')
-              .map((c: any) => c.text || '')
-              .join('\n')
-          : typeof contentArray === 'string'
-            ? contentArray
-            : '';
-        const displayText = this.extractDisplayUserContent(textContent);
-        if (displayText && displayText.trim()) {
-          messages.push({
-            id: `${conversationId}-${messageIndex++}`,
-            conversation_id: conversationId,
-            type: 'text',
-            role: 'user',
-            position: 'right',
-            content: { content: displayText },
-            create_time: timestamp,
-            status: finishedMessageStatus,
-          } as unknown as TMessage);
-        }
-        continue;
-      }
-
-      if (msgRole === 'assistant' && Array.isArray(contentArray)) {
-        for (const block of contentArray) {
-          if (block?.type === 'thinking') {
-            // thinking content is not displayed in UI
-          } else if (block?.type === 'text') {
-            const textContent = block.text || '';
-            if (textContent && textContent.trim()) {
-              messages.push({
-                id: `${conversationId}-${messageIndex++}`,
-                conversation_id: conversationId,
-                type: 'text',
-                role: 'assistant',
-                position: 'left',
-                content: { content: textContent },
-                create_time: timestamp,
-                status: finishedMessageStatus,
-              } as unknown as TMessage);
-            }
-          } else if (block?.type === 'tool_use') {
-            const toolCallId = block.id || `${conversationId}-${messageIndex}`;
-            const responseToolCallId = block.uuid || toolCallId;
-            const rawInput = this.parseMossToolInput(block.input);
-            if (block.name === 'AskUserQuestion') {
-              const question = typeof rawInput.question === 'string' ? rawInput.question : '';
-              const description = typeof rawInput.description === 'string' ? rawInput.description : undefined;
-              const options = Array.isArray(rawInput.options) ? rawInput.options.filter((option): option is string => typeof option === 'string') : [];
-
-              messages.push({
-                id: `${conversationId}-${messageIndex++}`,
-                msg_id: toolCallId,
-                conversation_id: conversationId,
-                type: 'acp_question',
-                position: 'left',
-                content: {
-                  question,
-                  intro: description,
-                  options,
-                  conversationId,
-                  toolCallId,
-                  responseToolCallId,
-                },
-                create_time: timestamp,
-                status: finishedMessageStatus,
-              } as unknown as TMessage);
-            } else {
-              messages.push({
-                id: `${conversationId}-${messageIndex++}`,
-                msg_id: toolCallId,
-                conversation_id: conversationId,
-                type: 'acp_tool_call',
-                position: 'left',
-                content: {
-                  sessionId: mossSessionId,
-                  update: {
-                    sessionUpdate: 'tool_call',
-                    toolCallId,
-                    status: 'completed',
-                    title: block.name,
-                    kind: this.getMossToolKind(block.name || ''),
-                    rawInput,
-                    content: [],
-                  },
-                },
-                create_time: timestamp,
-                status: finishedMessageStatus,
-              } as unknown as TMessage);
-            }
-          }
-        }
-      } else if (msgRole === 'assistant') {
-        const textContent = typeof contentArray === 'string' ? contentArray : '';
-        if (textContent && textContent.trim()) {
-          messages.push({
-            id: `${conversationId}-${messageIndex++}`,
-            conversation_id: conversationId,
-            type: 'text',
-            role: 'assistant',
-            position: 'left',
-            content: { content: textContent },
-            create_time: timestamp,
-            status: finishedMessageStatus,
-          } as unknown as TMessage);
-        }
-      }
-    }
-
-    return { messages, foundModel };
-  }
-
-  private parseMossToolInput(input: unknown): Record<string, unknown> {
-    if (!input) return {};
-    if (typeof input === 'string') {
-      try {
-        const parsed = JSON.parse(input);
-        return parsed && typeof parsed === 'object' ? (parsed as Record<string, unknown>) : {};
-      } catch {
-        return { input };
-      }
-    }
-    return typeof input === 'object' ? (input as Record<string, unknown>) : { input };
-  }
-
-  private extractDisplayUserContent(content: string): string {
-    let userContent = content;
-    if (userContent.includes('[User Request]')) {
-      const parts = userContent.split('[User Request]');
-      userContent = parts[parts.length - 1]?.trim() || userContent;
-    }
-    return userContent;
-  }
-
-  private getMossToolKind(toolName: string): 'read' | 'edit' | 'execute' {
-    const normalized = toolName.toLowerCase();
-    if (normalized.includes('read')) return 'read';
-    if (normalized.includes('write') || normalized.includes('edit') || normalized.includes('patch')) return 'edit';
-    return 'execute';
-  }
-
-  /**
    * Get messages - LOCAL-FIRST approach
    * 获取消息 - 本地优先策略
    *
@@ -798,7 +496,7 @@ export class RemoteConversationProvider implements IConversationProvider {
         }
 
         const allMessages = contextData.context.messages;
-        const { messages, foundModel } = this.convertMossMessagesToTMessages(allMessages, conversationId, mossSessionId);
+        const { messages, foundModel } = convertMossMessagesToTMessages(allMessages, conversationId, mossSessionId);
 
         // Save to local DB for future local-first reads / 保存到本地数据库供后续本地优先读取
         let insertedCount = 0;
@@ -883,7 +581,7 @@ export class RemoteConversationProvider implements IConversationProvider {
       }
 
       const allMessages = contextData.context.messages;
-      const { messages, foundModel } = this.convertMossMessagesToTMessages(allMessages, conversationId, mossSessionId);
+      const { messages, foundModel } = convertMossMessagesToTMessages(allMessages, conversationId, mossSessionId);
 
       // Clear existing local messages to avoid duplicates / 清除现有本地消息避免重复
       db.deleteConversationMessages(conversationId);
